+++ CourseApp/ICoder.cs
@@ -1,0 +1,7 @@
+﻿namespace CourseApp
+{
+    public interface ICoder
+    {
+        void Code();
+    }
+}
