+++ CourseApp/Programmer.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace CourseApp
+{
+    public class Programmer : Employee, ICoder
+    {
+        void ICoder.Code()
+        {
+            Console.WriteLine("Окей гугл, открой стак оверфлоу");
+        }
+    }
+}
